refactor: move isMyself to isValidPeer
