+++ packages/core-p2p/src/utils/index.ts
@@ -1,7 +1,5 @@
 export { checkDNS } from "./check-dns";
 export { checkNTP } from "./check-ntp";
-export { isMyself } from "./is-myself";
 export { isWhitelisted } from "./is-whitelisted";
-export { parseRemoteAddress } from "./parse-remote-address";
 export { restorePeers } from "./restore-peers";
 export { isValidPeer } from "./is-valid-peer";
