+++ packages/core-p2p/src/utils/is-valid-peer.ts
@@ -1,11 +1,13 @@
-import { isMyself } from "./is-myself";
-import { parseRemoteAddress } from "./parse-remote-address";
+import { process } from "ipaddr.js";
+import os from "os";
 
 /**
  * Checks if the peer is a valid remote peer.
  */
 export const isValidPeer = (peer: { ip: string; status?: string | number }): boolean => {
-    if (!parseRemoteAddress(peer)) {
+    peer.ip = sanitizeRemoteAddress(peer.ip);
+
+    if (!peer.ip) {
         return false;
     }
 
@@ -20,4 +22,25 @@ export const isValidPeer = (peer: { ip: string; status?: string | number }): boo
     }
 
     return true;
+};
+
+/**
+ * Sanitizes the given ip and returns it sanitized ip on success or null on failure.
+ */
+const sanitizeRemoteAddress = (ip: string): string | null => {
+    try {
+        return process(ip).toString();
+    } catch (error) {
+        return null;
+    }
+};
+
+const isMyself = (ipAddress: string): boolean => {
+    const interfaces = os.networkInterfaces();
+
+    return (
+        ipAddress.startsWith("127.") ||
+        ipAddress.startsWith("0.") ||
+        Object.keys(interfaces).some(ifname => interfaces[ifname].some(iface => iface.address === ipAddress))
+    );
 };
