--- packages/core-p2p/src/utils/parse-remote-address.ts
@@ -1,14 +1,0 @@
-import { process } from "ipaddr.js";
-
-/**
- * Parses the given ip and returns wether it could be parsed or not.
- */
-export const parseRemoteAddress = (peer: { ip: string }): boolean => {
-    try {
-        peer.ip = process(peer.ip).toString();
-
-        return true;
-    } catch (error) {
-        return false;
-    }
-};
